requirements.txt
certifi==2024.2.2
cmake==3.29.2
influxdb-client==1.42.0
paho-mqtt==2.0.0
pylsl==1.16.2
python-dateutil==2.9.0.post0
reactivex==4.0.4
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
